+++ benight-admin/src/app/app.module.ts
@@ -1,7 +1,7 @@
 
 import { NgModule } from '@angular/core'
 
-import { myAppProviders, myAppModules, FirebaseModules} from '@bn8-core/imports'
+import { myAppProviders, myAppModules, FirebaseModules, FirebaseProviders} from '@bn8-core/imports'
 
 import { AppComponent } from './app.component'
 import { AppRoutingModule } from './app-routing.module';
@@ -17,7 +17,8 @@ import { FireFormDirective } from './directives/fire-form.directive'
     AppRoutingModule
   ],
   providers: [
-    ...myAppProviders
+    ...myAppProviders,
+    ...FirebaseProviders
   ],
   bootstrap: [AppComponent]
 })