+++ benight-admin/src/app/services/db-extension.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DbExtensionService } from './db-extension.service';
+
+describe('DbExtensionService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: DbExtensionService = TestBed.get(DbExtensionService);
+    expect(service).toBeTruthy();
+  });
+});